add makestyles, move search to nonhero
