+++ src/components/SmallCourseCard/index.js
