--- src/pages/HomeView/courses/Recommended.module.css
@@ -1,14 +1,0 @@
-.container{
-    margin-bottom: 2rem;
-    padding: 1rem 0;
-    display: flex;
-    flex-wrap: nowrap;
-    overflow-x: auto;
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-}
-
-.mainHeading{
-    margin: 1rem 0 0 1rem !important;
-    font-size: 1rem !important;
-}
