+++ src/pages/HomeView/index.js
@@ -1,8 +1,6 @@
 import React from "react";
 import Hero from "./Hero";
 import NonHero from "./NonHero";
-import Recommended from "./courses/Recommended";
-import Search from "./Search";
 import { makeStyles } from "@material-ui/core";
 import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp";
 import Fab from "@material-ui/core/Fab";
@@ -25,8 +23,6 @@ export default function HomeView(props) {
       <Hero />
       <div id="back-to-top-anchor"></div>
       <NonHero />
-      <Search />
-      <Recommended />
       <div className={classes.backtotop}>
         <ScrollTop {...props}>
           <Fab color="secondary" size="small" aria-label="scroll back to top">
